Add files via upload
Complete project where the use of SQL database and html interface to find the perfect team for India, considering some criteria such as the location of the match to be held, opponent team, player experience, etc. It uses weight system to finally give out the rating to each player and then rank them according to it.